+++ src/views/layout.tsx
@@ -10,6 +10,7 @@ import {
   FORM_PERSIST_SCRIPT,
   EMAIL_VALIDATE_SCRIPT,
   PROOF_UPLOAD_SCRIPT,
+  STAGE_PROOF_SCRIPT,
 } from "./scripts";
 
 // Meta carries page-level SEO/social metadata into the Layout.
@@ -49,6 +50,7 @@ const GlobalScripts: FC = () => (
     <script dangerouslySetInnerHTML={{ __html: FORM_PERSIST_SCRIPT }} />
     <script dangerouslySetInnerHTML={{ __html: EMAIL_VALIDATE_SCRIPT }} />
     <script dangerouslySetInnerHTML={{ __html: PROOF_UPLOAD_SCRIPT }} />
+    <script dangerouslySetInnerHTML={{ __html: STAGE_PROOF_SCRIPT }} />
   </>
 );